feat(modal): added home and search wrapper
